Major backwards-incompatible configuration changes, bump version to 1.2.0
The config, pidfile, and logfile settings have had their defaults
changed so that they no longer include directories:

  Before:
    config - config/style.yaml
    logfile - log/style.log
    pidfile - log/style.pid
  After:
    config - style.yaml
    logfile - style.log
    pidfile - style.pid

The ramaze and generic adapters have been removed, for non-rails
installations, the adapter should be the name of a file to
require that will load the adapter/framework.  See the README for
upgrade instructions.

The RDocs are now hosted on RubyForge.

Add a Rakefile and .gitignore file for ease of maintenance.